Replace Errno with dedicated Error types.
Errno is no longer a subclass of Exception in Crystal, instead it is an
Enum and becomes more of an implementation detail.

This set of changes doesn't go as far as it could for detecting the
Errno and making appropriate types, but it does get things compiling
again on the latest version of Crystal.